Изменение в механизме тестирования
