+++ demoForm/src/test/java/com/example/demo/core/services/CatalogServiceTest.java
@@ -1,63 +1,77 @@
 package com.example.demo.core.services;
 
-import org.junit.jupiter.api.RepeatedTest;
+import com.example.demo.SearchParameterAnalyzer;
+import com.example.demo.core.repos.CatalogRepository;
 import org.junit.jupiter.api.Test;
-
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.data.domain.Pageable;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+@SpringBootTest
+class CatalogServiceTest {
 
-    class CatalogServiceTest {
+    @Autowired
+    CatalogService catalogService;
 
-        @RepeatedTest(10)
-        void testing () {
-            String[] cases = {"майка 123-321","123-321 майка","майка","123-321","123"};
-            String[] results = {
-                    "?search=майка&priceFrom=123&priceTo=321",
-                    "?search=майка&priceFrom=123&priceTo=321",
-                    "?search=майка&priceFrom=0&priceTo=999999",
-                    "?priceFrom=123&priceTo=321",
-                    "?priceFrom=123&priceTo=999999"
-            };
-            int generateRandomNumber = (int) ((Math.random() * (5-0)) + 0);
-            System.out.println(cases[generateRandomNumber]);
-            String searchParameter = cases[generateRandomNumber];
-            String actual = searchAnalysis(searchParameter);
-            String expected = results[generateRandomNumber];
-            assertEquals(actual, expected);
-        }
+    @Autowired
+    CatalogRepository catalogRepository;
 
-        String searchAnalysis(String search) {
-            String productName;
-            String priceFrom;
-            String priceTo;
-            Pattern pattern = Pattern.compile("([а-яА-Я]+)?(\\s+)?(\\d+)?(\\s+)?-?(\\s+)?(\\d+)?(\\s+)?([а-яА-Я]+)?");
-            Matcher matcher = pattern.matcher(search);
-            matcher.find();
-            productName = matcher.group(1);
-            if (productName == null) {
-                productName = matcher.group(8);
-            }
-            priceFrom = matcher.group(3);
-            if (priceFrom == null) {
-                priceFrom = "0";
-            }
-            priceTo = matcher.group(6);
-            if (priceTo == null) {
-                priceTo = "999999";
-            }
-            String result = "?";
-            if (!(productName == null) && (!productName.isEmpty() || !productName.isBlank())) {
-                result = result + "search=" + productName;
-            }
-            if (!(productName == null)) {
-                result = result + "&priceFrom=" + priceFrom;
-            } else {
-                result = result + "priceFrom=" + priceFrom;
-            }
-            result = result + "&priceTo=" + priceTo;
-            return result;
+    Pageable pageable;
+
+    SearchParameterAnalyzer searchParameterAnalyzer = new SearchParameterAnalyzer();
+
+    String[] cases = {
+            "майка 123-321",
+            "123-321 майка",
+            "майка",
+            "123-321",
+            "123"
+    };
+    String[][] expected = {
+            {"майка", "123.0", "321.0"},
+            {"майка", "123.0", "321.0"},
+            {"майка", "0.0", "9999.0"},
+            {"empty", "123.0", "321.0"},
+            {"empty", "123.0", "9999.0"}
+    };
+
+    @Test
+    public void testingSearchParameterAnalyzer() {
+        for (int i=0; i<cases.length; i++) {
+            searchParameterAnalyzer.analyzingSearchParameter(cases[i]);
+                if (searchParameterAnalyzer.getNameOfSearchedProduct() == null) {
+                    searchParameterAnalyzer.setNameOfSearchedProduct("empty");
+                }
+                    assertEquals(expected[i][0], searchParameterAnalyzer.getNameOfSearchedProduct(),
+                            "Should contain productname");
+                    assertEquals(expected[i][1], searchParameterAnalyzer.getBottomLineOfPriceInSearchedProduct().toString(),
+                            "Should contain firstprice");
+                    assertEquals(expected[i][2], searchParameterAnalyzer.getTopLineOfPriceInSearchedProduct().toString(),
+                            "Should contain secondprice");
         }
     }
+
+    String search = "майка";
+
+    @Test
+    public void testing() {
+        try {
+            var rez = catalogService.listAll(search, pageable);
+            assertAll(
+                    () -> assertEquals("майка", rez.getContent().get(0).getProductName()),
+                    () -> assertEquals(200, rez.getContent().get(0).getProductPrice())
+            );
+        } catch (Exception ex) {
+            System.out.println("Search parameter is empty, showing all catalog");
+            catalogRepository.findAll();
+        }
+    }
+
+
+
+    
+
+
+}
